pdf_tables/requirements.txt
absl-py==0.6.1
appnope==0.1.0
argcomplete==1.8.2
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.5.3
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
distro==1.3.0
docx2txt==0.6
EbookLib==0.15
en-core-web-sm==2.0.0
entrypoints==0.2.3
flake8==3.5.0
future==0.17.1
gast==0.2.0
Glances==2.11.1
grpcio==1.17.1
h5py==2.9.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.0.8
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
livereload==2.5.2
lxml==4.2.4
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.3
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.2.0
pathlib==1.0.1
pdfminer.six==20181108
pdftotree==0.4.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.2.0
plac==0.9.6
pluggy==0.7.1
pocketsphinx==0.1.15
preshed==1.0.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.6
ptyprocess==0.5.2
py==1.6.0
pycodestyle==2.3.1
pycryptodome==3.6.6
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pytest==3.8.1
python-dateutil==2.7.2
python-pptx==0.6.5
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.0.0
regex==2017.4.5
requests==2.19.1
scikit-image==0.13.1
scikit-learn==0.20.2
scipy==1.1.0
selectivesearch==0.4
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
sortedcontainers==2.1.0
spacy==2.0.12
SpeechRecognition==3.6.3
style==1.1.0
tabula-py==1.3.1
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
tesseract==0.1.3
testpath==0.3.1
thinc==6.10.3
toolz==0.9.0
tornado==5.0.2
tqdm==4.25.0
traitlets==4.3.2
ujson==1.35
update==0.0.1
urllib3==1.23
Wand==0.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==1.1.0
